+++ anyblok_wms_base/core/operation/move.py
@@ -43,19 +43,24 @@ class Move(Mixin.WmsSingleInputOperation,
 
     def after_insert(self):
         state, to_move, dt_exec = self.state, self.input, self.dt_execution
+        orig_dt_until = to_move.dt_until
+        if orig_dt_until is None:
+            dt_until = None
+        else:
+            # it's not clear what it means for to_move.dt_until not to be None
+            # in any case, dt_until should be at least dt_from
+            dt_until = max(orig_dt_until, dt_exec)
 
+        to_move.dt_until = dt_exec
+        if state == 'done':
+            to_move.state = 'past'
         self.registry.Wms.PhysObj.Avatar.insert(
             location=self.destination,
             outcome_of=self,
             state='present' if state == 'done' else 'future',
             dt_from=dt_exec,
-            # copied fields:
-            dt_until=to_move.dt_until,
+            dt_until=dt_until,
             obj=to_move.obj)
-
-        to_move.dt_until = dt_exec
-        if state == 'done':
-            to_move.state = 'past'
 
     @classmethod
     def check_create_conditions(cls, state, dt_execution, destination=None,
@@ -91,10 +96,8 @@ class Move(Mixin.WmsSingleInputOperation,
     def execute_planned(self):
         dt_execution = self.dt_execution
 
-        self.outcome.update(state='present', dt_from=dt_execution)
-        self.registry.flush()  # TODO should now been unneeded
-
         self.input.update(state='past', dt_until=dt_execution)
+        self.outcome.update(state='present', dt_from=dt_execution)
 
     def is_reversible(self):
         """Moves are always reversible.